+++ CHANGELOG.md
@@ -73,6 +73,7 @@
 ### ⬆️ Improved
 
 ### ✅ Added
+- Introduce **experimental** `DirectChannelInfoScreen` and `GroupChannelInfoScreen` for displaying channel information and the available channel actions. [#5773](https://github.com/GetStream/stream-chat-android/pull/5773)
 
 ### ⚠️ Changed